split cdrom into separate file
